Update for slider use case, refactor code

+++ Assets/Scripts/Audio.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 public abstract class Audio
 {
     [Range(0.0f, 1.0f)]
-    public int DefaultVolume = 1;
+    public float DefaultVolume = 1;
 
     [Range(0.0f, 1.0f)]
-    public int Volume = 1;
+    public float Volume = 1;
 
     public AudioClip Clip;
 }
